[DEV-200] convert nonce to uint64

+++ btcjson/dagsvrresults.go
@@ -38,7 +38,7 @@ type GetBlockVerboseResult struct {
 	Tx             []string      `json:"tx,omitempty"`
 	RawTx          []TxRawResult `json:"rawtx,omitempty"`
 	Time           int64         `json:"time"`
-	Nonce          uint32        `json:"nonce"`
+	Nonce          uint64        `json:"nonce"`
 	Bits           string        `json:"bits"`
 	Difficulty     float64       `json:"difficulty"`
 	PreviousHashes []string      `json:"previousblockhashes"`
